+++ open_instruct/environments/__init__.py
@@ -1,5 +1,6 @@
 """RL Environments for open-instruct."""
 
-from .base import EnvCall, RLEnvironment, StepResult
+from .base import BaseEnvConfig, EnvCall, RLEnvironment, RolloutState, StepResult
+from .pool import EnvironmentPool
 
-__all__ = ["EnvCall", "RLEnvironment", "StepResult"]
+__all__ = ["BaseEnvConfig", "EnvCall", "RLEnvironment", "StepResult", "RolloutState", "EnvironmentPool"]