+++ scripts/train/debug/tools/dr_tulu_parser_debug.sh
@@ -66,7 +66,7 @@ VLLM_ALLOW_INSECURE_SERIALIZATION=1 uv run --extra dr-tulu open_instruct/grpo_fa
     --tools dr_agent_mcp \
     --tool_parser dr_tulu \
     --tool_configs '{"tool_names": "snippet_search,google_search,browse_webpage", "parser_name": "v20250824", "host": "'"$MCP_HOST"'", "port": '"$MCP_PORT"'}' \
-    --max_tool_calls 5 \
+    --max_steps 5 \
     --pass_tools_to_chat_template false \
     --push_to_hub false